clean

+++ src/player/ai/nn/AFUtil.h
@@ -1,14 +1,14 @@
 #pragma once
 #include <cstdlib>
 
-double sigm(double x) {
+float sigm(float x) {
 	return 1 / (1 + exp(-x));
 }
 
-double tanh1(double x) {
+float tanh1(float x) {
 	return tanh(x);
 }
 
-double fsigm(double x) {
+float fsigm(float x) {
 	return x / (1 + abs(x));
 }
